travel-model/src/test/java/org/olebas/travel/app/service/GeographicServiceImplTest.java
package org.olebas.travel.app.service;

import org.junit.Before;
import org.junit.Test;
import org.olebas.travel.app.model.entity.geography.City;
import org.olebas.travel.app.service.impl.GeographicServiceImpl;

import java.util.List;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;

public class GeographicServiceImplTest {
    private GeographicService service;

    @Before
    public void setup() {
        service = new GeographicServiceImpl();
    }

    @Test
    public void testNoDataReturnedAtStart() {
        List<City> cities = service.findCities();
        assertTrue(cities.isEmpty());
    }

    @Test
    public void testSaveNewCitySuccess() {
        City city = new City("Kyiv");
        service.saveCity(city);

        List<City> cities = service.findCities();
        assertEquals(cities.size(), 1);
        assertEquals(cities.get(0).getName(), "Kyiv");
    }
}
